+++ examples/react-native/jest.config.js
@@ -3,7 +3,7 @@ const {resolve} = require('path');
 module.exports = {
   preset: 'react-native',
   testEnvironmentOptions: {
-    globalsCleanupMode: 'soft',
+    globalsCleanup: 'soft',
   },
   // this is specific to the Jest repo, not generally needed (the files we ignore will be in node_modules which is ignored by default)
   transformIgnorePatterns: [resolve(__dirname, '../../packages')],